refactor: rename store/ → state/ with split
Runtime atoms (liveStatus, liveUpdates, sourceContext) move to
state/runtime/. Config infrastructure (drafts, persist, reactions)
moves to state/ root next to state/settings/. The "config" prefix
drops from drafts/reactions filenames since their parent already
implies "settings infrastructure".

After this and the config→settings rename, all reactive state lives
under state/, with settings/ for persisted user-tunable and runtime/
for session-only.
