+++ app/src/scene/system/cameraRig.ts
@@ -27,7 +27,7 @@
 import * as THREE from 'three';
 import { OrbitControls } from 'three/addons/controls/OrbitControls.js';
 import { CAMERA_PERSPECTIVE, CAMERA_CONTROLS, ANIMATION_TIMING } from '@/state/settings/index.js';
-import { CURRENT_SOURCE_KEY } from '@/store/sourceContext.js';
+import { CURRENT_SOURCE_KEY } from '@/state/runtime/sourceContext.js';
 import { StreetAxis } from '@/types';
 import type { Building, Street } from '@/types';
 import type { createWorld } from '../world.js';
