+++ app/tests/views/panes/controlsPane.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, beforeAll } from 'vitest';
 import { buildControlsPane } from '@/views/panes/controlsPane.js';
-import { attachPersistence } from '@/store/persist.js';
+import { attachPersistence } from '@/state/persist.js';
 import * as Config from '@/state/settings/index.js';
 
 describe('buildControlsPane', () => {
